Create LICENSE file
The license is specified as [ISC](https://choosealicense.com/licenses/isc/) in the package.json. This commit adds a traditional LICENSE file in the root of the repo, where many automated license checking tools expect to find it.

This also allows Github to find the license and indicate the repo's license in various places.